+++ aspnetcore/Entra.EventHandlers.AspNetCore.IntegrationTests/Endpoints/AttributeCollectionSubmitEndpointTests.cs
@@ -1,6 +1,6 @@
 ﻿using Entra.EventHandlers.AspNetCore.IntegrationTests.Utils;
 using Entra.EventHandlers.AspNetCore.IntegrationTests.Utils.AppFactories;
-using Entra.EventHandlers.TestHelpers;
+using Entra.EventHandlers.TestData;
 using FluentAssertions;
 using System.Net;
 using System.Text;