+++ aspnetcore/Entra.EventHandlers.AspNetCore.IntegrationTests/Utils/TestPasswordSubmitHandlerThrows.cs
@@ -2,12 +2,13 @@
 using Entra.EventHandlers.Abstractions.Events;
 using Entra.EventHandlers.Abstractions.Interfaces;
 using Entra.EventHandlers.Abstractions.Responses;
+using Entra.EventHandlers.Abstractions.Results;
 
 namespace Entra.EventHandlers.AspNetCore.IntegrationTests.Utils;
 
 public class TestPasswordSubmitHandlerThrows : IPasswordSubmitHandler
 {
-    public Task<PasswordSubmitResponse> HandleAsync(
+    public Task<EntraHandlerResult<PasswordSubmitResponse>> HandleAsync(
         PasswordSubmitEvent request,
         CancellationToken cancellationToken = default)
     {