+++ aspnetcore/Entra.EventHandlers.AspNetCore.IntegrationTests/Utils/TestVerifiedIdClaimValidationHandler.cs
@@ -1,6 +1,7 @@
 ﻿using Entra.EventHandlers.Abstractions.Events;
 using Entra.EventHandlers.Abstractions.Interfaces;
 using Entra.EventHandlers.Abstractions.Responses;
+using Entra.EventHandlers.Abstractions.Results;
 using Entra.EventHandlers.TestHelpers;
 using Entra.EventHandlers.Workforce.Builders;
 
@@ -8,7 +9,7 @@ namespace Entra.EventHandlers.AspNetCore.IntegrationTests.Utils;
 
 public class TestVerifiedIdClaimValidationHandler : TestHandlerBase, IVerifiedIdClaimValidationHandler
 {
-    public Task<VerifiedIdClaimValidationResponse> HandleAsync(
+    public Task<EntraHandlerResult<VerifiedIdClaimValidationResponse>> HandleAsync(
         VerifiedIdClaimValidationEvent request,
         CancellationToken cancellationToken = default)
     {
@@ -17,10 +18,10 @@ public class TestVerifiedIdClaimValidationHandler : TestHandlerBase, IVerifiedId
         WasCalled = true;
         CapturedCancellationToken = cancellationToken;
 
-        return Task.FromResult(
-            EntraWorkforceEventResponses
-                .VerifiedIdClaimValidation()
-                .Pass()
-                .Build());
+        var response = EntraWorkforceEventResponses.VerifiedIdClaimValidation()
+            .Pass()
+            .Build();
+
+        return Task.FromResult(new EntraHandlerResult<VerifiedIdClaimValidationResponse>(response));
     }
 }