+++ aspnetcore/Entra.EventHandlers.AspNetCore.UnitTests/Abstractions/TestEntraExceptionHandler.cs
@@ -1,5 +1,4 @@
 ﻿using Entra.EventHandlers.AspNetCore.Interfaces;
-using Microsoft.AspNetCore.Http;
 
 namespace Entra.EventHandlers.AspNetCore.UnitTests.Abstractions;
 
@@ -7,7 +6,7 @@ internal sealed class TestEntraExceptionHandler : IEntraExceptionHandler
 {
     public bool WasCalled { get; private set; }
 
-    public Task HandleAsync(Exception ex, HttpContext context, bool isEntraException)
+    public Task HandleAsync(Exception ex)
     {
         WasCalled = true;
         return Task.CompletedTask;