+++ aspnetcore/Entra.EventHandlers.AspNetCore/Adapters/ResponseAdapter.cs
@@ -1,5 +1,6 @@
 ﻿using Entra.EventHandlers.Abstractions.Errors;
 using Entra.EventHandlers.Abstractions.Responses;
+using Entra.EventHandlers.Hosting.Errors;
 using System.Text.Json;
 
 namespace Entra.EventHandlers.AspNetCore.Adapters;