+++ core/Entra.EventHandlers/Handlers/Base/TokenIssuanceStartHandlerBase.cs
@@ -1,6 +1,8 @@
 ﻿using Entra.EventHandlers.Abstractions.Events;
+using Entra.EventHandlers.Abstractions.Extensions;
 using Entra.EventHandlers.Abstractions.Interfaces;
 using Entra.EventHandlers.Abstractions.Responses;
+using Entra.EventHandlers.Abstractions.Results;
 using Entra.EventHandlers.Builders;
 using Microsoft.Extensions.Logging;
 using System.Diagnostics;
@@ -27,7 +29,7 @@ public abstract class TokenIssuanceStartHandlerBase(ILogger logger) : ITokenIssu
     /// In case of failure, a valid response containing an empty claim set is
     /// returned to ensure token issuance continues without interruption.
     /// </remarks>
-    public async Task<TokenIssuanceStartResponse> HandleAsync(TokenIssuanceStartEvent request, CancellationToken cancellationToken)
+    public async Task<EntraHandlerResult<TokenIssuanceStartResponse>> HandleAsync(TokenIssuanceStartEvent request, CancellationToken cancellationToken = default)
     {
         using var scope = Logger.BeginScope(new Dictionary<string, object?>
         {
@@ -38,7 +40,7 @@ public abstract class TokenIssuanceStartHandlerBase(ILogger logger) : ITokenIssu
 
         var sw = Stopwatch.StartNew();
 
-        Logger.LogInformation("Handling event");
+        Logger.LogInformation("Starting Entra event handling.");
 
         try
         {
@@ -51,25 +53,36 @@ public abstract class TokenIssuanceStartHandlerBase(ILogger logger) : ITokenIssu
             var actionType = response.Data.Actions.FirstOrDefault()?.OdataType ?? "None";
 
             Logger.LogInformation(
-                "Successfully handled event. DurationMs={Duration}, Action={ActionType}",
+                "Entra event handled successfully. DurationMs={DurationMs}, Action={ActionType}.",
                 sw.ElapsedMilliseconds,
                 actionType);
 
-            return response!;
+            return new EntraHandlerResult<TokenIssuanceStartResponse>(response);
         }
         catch (Exception ex)
         {
             sw.Stop();
 
-            Logger.LogError(
-                ex,
-                "Unhandled exception. DurationMs={Duration}",
-                sw.ElapsedMilliseconds);
+            if (ex.IsEntraException())
+            {
+                Logger.LogWarning(
+                    ex,
+                    "Entra domain exception occurred during Entra event handling. DurationMs={DurationMs}.",
+                    sw.ElapsedMilliseconds);
+            }
+            else
+            {
+                Logger.LogError(
+                    ex,
+                    "Unexpected failure occurred during Entra event handling. DurationMs={DurationMs}.",
+                    sw.ElapsedMilliseconds);
+            }
 
-            return EntraEventResponses
-                .TokenIssuanceStart()
+            var defaultResponse = EntraEventResponses.TokenIssuanceStart()
                 .ProvideClaimsForToken([])
                 .Build();
+
+            return new EntraHandlerResult<TokenIssuanceStartResponse>(defaultResponse, ex);
         }
     }
 
@@ -78,5 +91,5 @@ public abstract class TokenIssuanceStartHandlerBase(ILogger logger) : ITokenIssu
     /// TokenIssuanceStart event. Implementations should override
     /// this method instead of <see cref="HandleAsync"/>.
     /// </summary>
-    protected abstract Task<TokenIssuanceStartResponse> HandleCoreAsync(TokenIssuanceStartEvent request, CancellationToken cancellationToken);
+    protected abstract Task<TokenIssuanceStartResponse> HandleCoreAsync(TokenIssuanceStartEvent request, CancellationToken cancellationToken = default);
 }